test: split provider-failure suite and align transient-retry timers/copy
Fix lint max-lines by moving the connection/timeout transient-retry cases
into agent-runner-execution-transient-retry.test.ts, drop the now-unused
FallbackRunnerParams import, and change the catch alias to const.

Update the CLI/Codex timeout assertions to the current surfaced copy and
advance the retry backoff by a bounded interval instead of draining all
timers so the one-shot orchestrator retry settles deterministically.
